Allow for larger values for resilience feature using unsigned short instead of unsigned byte.

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -3774,7 +3774,7 @@ void ObjectMgr::LoadPlayerInfo()
                     levelInfo.stats[i] = fields[i + 3].GetUInt8();
 
                 // Franca: Resilience Feature [Origin]
-                levelInfo.resilience = fields[8].GetUInt8();
+                levelInfo.resilience = fields[8].GetUInt16();
                 
             }
